ignore_sections parsing options
